Rename product-filter-stock-status to product-filter-status
