+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/constants.ts
@@ -8,7 +8,7 @@ export const EXCLUDED_BLOCKS = [
 	'woocommerce/product-filters',
 	'woocommerce/product-filter-attribute',
 	'woocommerce/product-filter-price',
-	'woocommerce/product-filter-stock-status',
+	'woocommerce/product-filter-status',
 	'woocommerce/product-collection',
 	'core/query',
 ];
