+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/checkbox-list/block.json
@@ -12,7 +12,7 @@
 	"apiVersion": 3,
 	"ancestor": [
 		"woocommerce/product-filter-attribute",
-		"woocommerce/product-filter-stock-status"
+		"woocommerce/product-filter-status"
 	],
 	"supports": {
 		"color": {
