+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/status-filter/inspector.tsx
@@ -34,9 +34,7 @@ export const Inspector = ( {
 
 	if ( displayStyleOptions.length === 0 ) {
 		displayStyleOptions = getBlockTypes().filter( ( blockType ) =>
-			blockType.ancestor?.includes(
-				'woocommerce/product-filter-stock-status'
-			)
+			blockType.ancestor?.includes( 'woocommerce/product-filter-status' )
 		);
 	}
