+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/status-filter/preview.tsx
