+++ plugins/woocommerce-blocks/assets/js/icons/library/product-filter-status.tsx
