+++ plugins/woocommerce-blocks/tests/e2e/content-templates/template_archive-product_stock-status.handlebars
@@ -10,7 +10,7 @@
 <!-- /wp:buttons -->
 <!-- /wp:woocommerce/product-filter-clear-button -->
 
-{{#> wp-block blockName='woocommerce/product-filter-stock-status' attributes=attributes }}
+{{#> wp-block blockName='woocommerce/product-filter-status' attributes=attributes }}
 {{/ wp-block }}
 
 <!-- wp:woocommerce/product-collection {"query":{"perPage":9,"pages":0,"offset":0,"postType":"product","order":"asc","orderBy":"title","search":"","exclude":[],"inherit":true,"taxQuery":[],"isProductCollectionBlock":true,"featured":false,"woocommerceOnSale":false,"woocommerceStockStatus":["instock","outofstock","onbackorder"],"woocommerceAttributes":[],"woocommerceHandPickedProducts":[]},"tagName":"div","displayLayout":{"type":"flex","columns":3,"shrinkColumns":true},"queryContextIncludes":["collection"]} -->
