+++ plugins/woocommerce/templates/parts/product-filters.html
@@ -17,8 +17,8 @@
 		<!-- /wp:button -->
 	</div>
 	<!-- /wp:buttons -->
-	<!-- wp:woocommerce/product-filter-stock-status -->
-	<div class="wp-block-woocommerce-product-filter-stock-status">
+	<!-- wp:woocommerce/product-filter-status -->
+	<div class="wp-block-woocommerce-product-filter-status">
 		<!-- wp:group {"metadata":{"name":"Header"},"style":{"spacing":{"blockGap":"0"}},"layout":{"type":"flex","flexWrap":"nowrap"}} -->
 		<div class="wp-block-group"><!-- wp:heading {"level":3} -->
 			<h3 class="wp-block-heading">Status</h3>
@@ -42,7 +42,6 @@
 		<div class="wp-block-woocommerce-product-filter-checkbox-list wc-block-product-filter-checkbox-list"></div>
 		<!-- /wp:woocommerce/product-filter-checkbox-list -->
 	</div>
-	<!-- /wp:woocommerce/product-filter-stock-status -->
+	<!-- /wp:woocommerce/product-filter-status -->
 </div>
 <!-- /wp:woocommerce/product-filters -->
-
